new dive list: better handling of selection change
This still has an odd bug. When we swipe the dive details to change the
selected dive and that action causes a trip to expand or collapse, then
the color change / highlight in the dive list isn't updated correctly,
even though the log file indicates that we made the necessary changes.
I think that may be because that change ends up changing the index
values - but this part needs more work.

But I consider this progress.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>