MobileSwipeModel: ignore changes of trips
We don't display them anyway, so reacting for change signals makes
no sense. We don't support changes to multiple trips or a mix
of trips and dives.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>

MobileSwipeModel: bug fixes

- add topLevelRowCountInSource() function
When adding items to the top-level, we have to count dives
as one entry and trips as the number of dives that it
contains.

This was done correctly when initializing the model but
not when adding top-level entries. Therefore, factor out
this code in its own function.

- Introduce *ForInsert functions to determine insertion indexes
This is not tested, but looks right. Testing will have to be done by:
1) inside of trip, trip stays in place
2) inside of trip, trip moves
3) outside of trip

To my knowledge, the base model does not yet(?) use move semantics
for moving from inside to outside of trip and vice-versa.

- Introduce the conept of "ranges" and convert full ranges.
This is needed for example when removing an empty trip, which turns
into a no-op. And, as we know by now, Qt model/view system doesn't
handle empty ranges gracefully(!).

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>