MobileListModel: fixes
1) expand trip if trip with dive is expanded

2) expand trip if we insert the current dive to trip

3) collapse trip if current dive is removed

4) fix mapRangeFromSourceForInsert()

Since we turn the source model upside down, we have to treat
indexes for insertion differently than indexes for access /
removal. Insertion indexes point to the item *before* which
we should insert. However, on reversal of the model that
becomes the item *after* which we insert.

The mapRangeFromSourceForInsert() did that by adding 1 to
the row where we insert. This didn't account for the fact
when determining the index we have to add the expanded trip:

Suppose we add before the expanded trip in the source model. The
mapRowFromSourceLevel() function turns things around and now
points to the trip entry. Since this is not past the trip,
we don't add the number of elements in the trip. Finally,
we add one to this index and now insert *into* the trip.

What we have to do instead is add one to the index before
taking into account the contents of the trip. This makes
us point past the trip and we correctly add the trip
contents. Thus, we get the correct insertion position.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>